requirements.txt
Flask==2.0.1
pymongo==3.11.4
python-dotenv==0.19.0
Flask-WTF==0.15.1
Werkzeug==2.2.2